Refactor: Pull out Newsletter components
